requirements.txt
vllm
ollama
torch
transformers
datasets
trl
pandas
numpy
scikit-learn
scipy
tqdm
PyYAML
openai